requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
bs64==0.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
EditorConfig==0.12.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.8
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
ruff==0.8.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
